Added Mafia class which implements IGame interface
Added the implemented methods from IGame class. Part of the handleMessage method in Mafia class has been done.

added ready, voteStart and gameStart methods. In these methods the ArrayList that are being used get converted into arrays. Would be easier to change the server privateMessage method to take in type ArrayList rather than array.

Now that the Server class has a privateMessage method for one recipient, parts of the handleMessage method in the Mafia class should be changed to make use of this.

changed variable "Origin" in IGame interface to origin.

Ian