Added elimDay, elimNight, checkElim and eliminate methods.
In eliminateDay method need to add check to see if the player has already voted to save the person who they are now voting to eliminate.

Realised that there is a potential deadlock in the game when using the old rules of the mafia needing to kill all the innocents. Haven't decided on how will handle this possible deadlock, however this might involve changing the win condition for the mafia to having equal to or greater than the number of innocents.