Added checkDay, checkElimNight, checkNight, save, voteDay and voteNight and elimNight methods.
Javadoc comments are missing for most of the methods.

Need to change vote methods to make use of the new boolean variable that is true if a vote is in progress (easier way to check if this is the case). Currently a vote could be started to change from day to night while there is a vote to eliminate a player.

Considering replacing a lot of the ArrayLists with maps instead. For example the id and player usernames could be kept in a map.