Changed all the ArrayLists back to being type Integer. So these ArrayLists are back to storing the userIDs of the players rather than their usernames. The variable that stored the username of the player who was on trial (being voted for in the vote to eliminate a player during the day) has been changed to store the userID of the player.
Changed the handleMessage method to first check if the game is in day or night phase. This means the amount of commands that the handleMessage needs to check for is greatly reduced for when it night (where there is only one acceptable command). This also means that the other methods no longer need to check whether it is day or night.

Changed the voteDay method so that it now no longer allows a player to vote multiple times to keep it day.

Added javadoc comments to the first two methods. Most of the methods are still missing comments.

Added org.apache.commons.collections4 to reference libraries. This is to make use of a Bidimap to store the id and names of players and use one to look up the other.