Added boolean variables to track if there is a vote for night in progress or a vote to eliminate a player (during the day). Implemented the TreeBidiMap in order to handle sometimes needing to use a players username name to look up their id as well as vice versa.
Added checkWin method that checks if either the innocents or the mafia win. Currently only sends out a message when either one wins. The rules for the mafia winning have been changed from the original implementation where the mafia win if they have the same number as the innocents rather than all innocents needing to be eliminated. This was because of realising a possible deadlock with the original game rules, where there could be one innocent and one mafia left and neither could get majority vote to beat the other. 

Added an assignPoints method as a placeholder.