Added check in start method for mafia which now prevents players who haven't readied up, voting to start the game.
Added a check in the ready method to make sure that the max number of players doesn't go over 15, there have also been two server messages added. A private message to inform a player that the Mafia game is already full and a public message to say when the game is full.

Added synchronized to every method in the Mafia class in order to try and prevent possible issues with racing between the different threads.