Added Timer and TimeTask to elimDayVote method. Unsure as to whether this will run properly (separately to the method so that the method doesn't wait for it to finished).
Added dayElimVoteTimeout method which is synchronised. This method is called by the timer after the set delay has occurred. This method checks if out of the current votes if the elimDay has majority if so it calls the elminateDay method, if not it calls the saved method.

Added checkSave method that checks to see if a majority has been reached whenever a save vote is made.

Added saved method that now ends the vote to eliminate a player if the majority of votes are reached to save them. 

Added cancelling of the dayElimTimer in the saved and elimDay methods.