Changes made to Mafia class
Added string array that takes in the starting mafia player names at the beginning of the game. This is then used at the end of the game to state which players were mafia.

Added a gameInProgress boolean variable that is currently a placeholder. Thinking of using this for te handleMessage method so that it is first checked and if false the only valid commands would be /ready, /unready and /start.