Mafia class removed  assignPoints method (placeholder anyway)
added MafiaTest class, this was to junit the methods in the Mafia class however apparently junit testing cannot be used this way. This class is being kept for the time being however should be deleted later on.

ServerThread class added, this was part of the attempt to get junit testing for the Mafia class, as above this is also being kept in with for the time being and should be deleted later on.