Mafia class
Mafia class now extends game

Constructor has been added to the Mafia class.

handleTimerEvent method has been added because of the game class. This method has not been used. 

instances where the origin was being removed from the ArrayList (as with ready, start etc) have been fixed to remove the index of the origin rather than try and use the origin id as the index.