Mafia class removed handleTimerEvent  method.
renamed the voting methods that didn't have vote in their name to include this for consistency. These were save and elimDay which are now saveVote and elimDayVote respectively.

for voting methods (elimDayVote, saveVote, elimNightVote) the first thing that is checked is whether the user being voted for is a player (still) in the game.

MafiaTest class
added methods to quickly start a 6 player game, a 8 player game(3 Mafia) .
added methods to end a 6player game by either innocent or mafia victory.
added method for a successful night vote for a 6 player game
added method for a successful night vote for a 3mafia (8 player) game
added tests to check the number of mafia for a game is correct as well as a test to make sure that the mafia selected are randomly chosen.

added tests for the dayVote method
added tests for the elimDayVote method
added tests for the nightVote method
added tests for the elimNightVote method
added tests for the saveVote method