Mafia class added dayTimersLengthMs as well as nightTimerLengthMs which are used for the length of these timers. Setter methods for these have been added so that JUnit tests could be sped up.
Renamed mafia ArrayList<Integer> to mafiaIDs to better represent the values being stored.

Changed checkElimNight to not take in a parameter.

Also made further fixes to handleLogout method. Added extra check to this method so when the leaver is a mafia player and this happened during the mafia night elimination vote then it is checked if the votes already cast are now all the votes of the mafia.


Mafia test
added testLengthTimers and testWaitLength variables which are faster version of the timers used in the normal Mafia class in order to speed up the game.  Added about 10 extra tests which test different situations for when a player logouts during different parts of the game.